requirements.txt
streamlit
selenium
requests
oauth2client
spacy
transformers
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
openai
colorama